// src/shared/types/titles.enum.ts
export enum TitleEnum {
  // Обов’язкові титули за квести (ігрові)
  CHESS_KNIGHT = 'CHESS_KNIGHT',
  QUEST_PRINCESS = 'QUEST_PRINCESS',
  PUZZLE_MASTER = 'PUZZLE_MASTER',
  WAY_LORD = 'WAY_LORD',
  MAGICIAN = 'MAGICIAN',
  GARDEN_KEEPER = 'GARDEN_KEEPER',
  WEDDING_PROPHET = 'WEDDING_PROPHET',

  // Додаткові ігрові (можуть бути видані за особливі умови)
  PHOENIX = 'PHOENIX',
  UNBROKEN = 'UNBROKEN',
  LOVE_WARRIOR = 'LOVE_WARRIOR',

  // Магазин
  ROMANCE_OVERLORD = 'ROMANCE_OVERLORD',
  ROMANCE_MAGE = 'ROMANCE_MAGE',
  HEARTBEAT_OF_LOVE = 'HEARTBEAT_OF_LOVE',
  ADVENTURE_MASTER = 'ADVENTURE_MASTER',
  LIGHT_HEART = 'LIGHT_HEART',
  SOUL_GUARDIAN = 'SOUL_GUARDIAN',
  SUNNY_BEAM = 'SUNNY_BEAM',
  DREAM_WEAVER = 'DREAM_WEAVER',
  MEMORY_ARCHITECT = 'MEMORY_ARCHITECT',
  TALE_HERO = 'TALE_HERO',
}